backend/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.23.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Validation
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# CORS
starlette==0.27.0

# Testing
pytest==7.4.3
httpx==0.25.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
